Keep RC-modified settings across modulator restarts

+++ src/ConfigParser.h
@@ -80,7 +80,7 @@ struct mod_settings_t {
     float cfrErrorClip = 1.0f;
 
     // Settings for the OFDM windowing
-    unsigned ofdmWindowOverlap = 0;
+    size_t ofdmWindowOverlap = 0;
 
 #if defined(HAVE_OUTPUT_UHD) || defined(HAVE_SOAPYSDR)
     Output::SDRDeviceConfig sdr_device_config;
